moving reload to stand-alone endpoint
(cherry picked from commit 7c54fd304fb61a74fc77b6b11e9978440d92517e)
