+++ descriptors/ModuleDescriptor-template.json
@@ -615,7 +615,7 @@
           "methods": [
             "POST"
           ],
-          "pathPattern": "/circulation/rules/",
+          "pathPattern": "/circulation/rules-reload",
           "modulePermissions": [
             "circulation.rules.get"
           ],
